+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Upcoming
+# v0.9.2
 
 ## Fixes